refactor(ui): complete tauri service domain split
