+++ src/ui/src/services/tauri/soundPacks.ts
@@ -1,1 +1,39 @@
-export {};
+import { invoke } from "@tauri-apps/api/core";
+import type { InstalledSoundPack, RegistryPack } from "../../types/soundpacks";
+
+export function cespFetchRegistry(): Promise<RegistryPack[]> {
+  return invoke("cesp_fetch_registry");
+}
+
+export function cespListInstalled(): Promise<InstalledSoundPack[]> {
+  return invoke("cesp_list_installed");
+}
+
+export function cespInstallPack(
+  name: string,
+  sourceRepo: string,
+  sourceRef: string,
+  sourcePath: string,
+): Promise<InstalledSoundPack> {
+  return invoke("cesp_install_pack", { name, sourceRepo, sourceRef, sourcePath });
+}
+
+export function cespUpdatePack(
+  name: string,
+  sourceRepo: string,
+  sourceRef: string,
+  sourcePath: string,
+): Promise<InstalledSoundPack> {
+  return invoke("cesp_update_pack", { name, sourceRepo, sourceRef, sourcePath });
+}
+
+export function cespDeletePack(name: string): Promise<void> {
+  return invoke("cesp_delete_pack", { name });
+}
+
+export function cespPreviewSound(
+  packName: string,
+  category: string,
+): Promise<void> {
+  return invoke("cesp_preview_sound", { packName, category });
+}
